requirements.txt
dnspython==2.2.1
hachoir==3.2.0
humanize==4.4.0
Pillow==9.3.0
pyaes==1.6.1
pymongo==4.3.3
Pyrogram==1.4.16
PySocks==1.7.1
python-dotenv==0.21.0
TgCrypto==1.2.5
